removed duplicate classes, organized classes
