+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
         </activity>
 
         <activity
-            android:name=".ShowProfile"
+            android:name=".users.ShowProfile"
             android:exported="true">
         </activity>
         <activity
