+++ app/src/main/java/com/example/nachosbusiness/admin_browse/ProfileDetailFragment.java
@@ -2,14 +2,11 @@ package com.example.nachosbusiness.admin_browse;
 
 import static android.view.View.GONE;
 
-import static com.google.firebase.appcheck.internal.util.Logger.TAG;
-
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -22,9 +19,7 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
-import com.example.nachosbusiness.DBManager;
 import com.example.nachosbusiness.R;
-import com.example.nachosbusiness.ShowProfile;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
