+++ app/src/main/java/com/example/nachosbusiness/events/Event.java
@@ -1,12 +1,11 @@
 package com.example.nachosbusiness.events;
 
-import com.example.nachosbusiness.QRUtil;
+import com.example.nachosbusiness.utils.QRUtil;
 import com.example.nachosbusiness.facilities.Facility;
 import com.google.firebase.Timestamp;
 
 import java.io.Serializable;
 import java.util.Date;
-import java.util.UUID;
 
 public class Event implements Serializable {
     private String eventID;
